improve enum conversions

+++ src/draw.rs
@@ -54,7 +54,7 @@ pub fn draw_map<P: AsRef<Path>>(path: P, map: Map, width: u8, height: u8, tiles:
     let mut image: RgbaImage = ImageBuffer::new(width as u32 * 16, height as u32 * 16);
 
     let is_glitch = matches!(map, Map::Glitch);
-    let bright_map = BRIGHT_MAPS.contains(&u8::from(map));
+    let bright_map = BRIGHT_MAPS.contains(&(map as u8));
 
     for (y, row) in tiles.into_iter().enumerate() {
         for (x, tile) in row.into_iter().enumerate() {
@@ -83,11 +83,11 @@ pub fn draw_map<P: AsRef<Path>>(path: P, map: Map, width: u8, height: u8, tiles:
                 if let Some(sprite) = &sprites[y][x] {
                     let sprite_index = match sprite {
                         Sprite::Collectible(collectible) => match collectible {
-                            Collectible::Sword | Collectible::SilverKey if bright_map => u8::from(*collectible) as usize + 18,
+                            Collectible::Sword | Collectible::SilverKey if bright_map => *collectible as usize + 18,
                             Collectible::PossumCoin => 100,
-                            _ => u8::from(*collectible) as usize + 17,
+                            _ => *collectible as usize + 17,
                         },
-                        Sprite::Door(door) => u8::from(*door) as usize + 2,
+                        Sprite::Door(door) => *door as usize + 2,
                         Sprite::WindRoute => 67,
                         Sprite::Save => 0,
                         Sprite::Other(_) => usize::MAX,
